# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * M-ALDBANI platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Project {
  id: string;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  category: string;
  technologies: string[];
  image: string;
  images?: string[];
  /** @nullable */
  results?: string | null;
  featured: boolean;
  order: number;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  category: string;
  technologies: string[];
  image: string;
  results?: string;
  featured?: boolean;
  order?: number;
}

export interface ProjectUpdate {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  category?: string;
  technologies?: string[];
  image?: string;
  results?: string;
  featured?: boolean;
  order?: number;
}

export interface Service {
  id: string;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  icon: string;
  /** @nullable */
  price?: string | null;
  featured: boolean;
  order: number;
}

export interface ServiceInput {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  icon: string;
  price?: string;
  featured?: boolean;
  order?: number;
}

export interface ServiceUpdate {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  icon?: string;
  price?: string;
  featured?: boolean;
  order?: number;
}

export interface Article {
  id: string;
  title: string;
  titleAr: string;
  slug: string;
  excerpt: string;
  excerptAr: string;
  content: string;
  contentAr: string;
  category: string;
  coverImage: string;
  readTime: number;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}

export interface ArticleInput {
  title: string;
  titleAr: string;
  slug: string;
  excerpt: string;
  excerptAr: string;
  content: string;
  contentAr: string;
  category: string;
  coverImage: string;
  readTime?: number;
  published?: boolean;
}

export interface ArticleUpdate {
  title?: string;
  titleAr?: string;
  excerpt?: string;
  excerptAr?: string;
  content?: string;
  contentAr?: string;
  category?: string;
  coverImage?: string;
  readTime?: number;
  published?: boolean;
}

export interface Consultation {
  id: string;
  /** @nullable */
  clientId?: string | null;
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  type: string;
  date: string;
  time: string;
  duration: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  /** @nullable */
  meetingLink?: string | null;
  /** @nullable */
  price?: number | null;
  paid: boolean;
  createdAt: string;
}

export interface BookingInput {
  clientName: string;
  clientEmail: string;
  clientPhone: string;
  type: string;
  date: string;
  time: string;
  duration: number;
  notes?: string;
}

export interface ConsultationUpdate {
  status?: string;
  adminNotes?: string;
  meetingLink?: string;
  price?: number;
  paid?: boolean;
}

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  message: string;
  subject?: string;
}

export interface Lead {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  source: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  value?: number | null;
  createdAt: string;
}

export interface LeadInput {
  name: string;
  email: string;
  phone?: string;
  source: string;
  status?: string;
  notes?: string;
  value?: number;
}

export interface LeadUpdate {
  name?: string;
  email?: string;
  phone?: string;
  source?: string;
  status?: string;
  notes?: string;
  value?: number;
}

export interface Message {
  id: string;
  senderId: string;
  senderName: string;
  content: string;
  read: boolean;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface SharedFile {
  id: string;
  name: string;
  url: string;
  size: number;
  type: string;
  uploadedAt: string;
}

export interface InvoiceItem {
  description: string;
  quantity: number;
  price: number;
}

export interface Invoice {
  id: string;
  number: string;
  amount: number;
  currency: string;
  status: string;
  dueDate: string;
  /** @nullable */
  paidAt?: string | null;
  items?: InvoiceItem[];
  createdAt: string;
}

export interface Notification {
  id: string;
  title: string;
  message: string;
  type: string;
  read: boolean;
  createdAt: string;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
}

export interface CountByType {
  type: string;
  count: number;
}

export interface AdminStats {
  totalVisitors: number;
  totalClients: number;
  totalConsultations: number;
  totalRevenue: number;
  pendingConsultations: number;
  newLeads: number;
  monthlyRevenue: MonthlyRevenue[];
  consultationsByType: CountByType[];
  leadsByStatus: CountByType[];
}

export interface PageStat {
  path: string;
  views: number;
}

export interface DailyVisitors {
  date: string;
  visitors: number;
}

export interface Analytics {
  pageViews: number;
  uniqueVisitors: number;
  avgSessionDuration: number;
  bounceRate: number;
  topPages: PageStat[];
  trafficSources: CountByType[];
  dailyVisitors: DailyVisitors[];
}

export interface Reaction {
  emoji: string;
  count: number;
  userReacted: boolean;
}

export interface Comment {
  id: string;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface CommunityPost {
  id: string;
  title: string;
  content: string;
  type: string;
  /** @nullable */
  image?: string | null;
  reactions: Reaction[];
  comments: Comment[];
  seenCount: number;
  publishedAt: string;
}

export interface CommunityPostInput {
  title: string;
  content: string;
  type: string;
  image?: string;
}

export interface ReactionInput {
  emoji: string;
}

export interface CommentInput {
  content: string;
}

export type ListProjectsParams = {
category?: string;
};

export type ListArticlesParams = {
category?: string;
limit?: number;
};

export type GetAvailableSlotsParams = {
date: string;
};

export type ListLeadsParams = {
status?: string;
};

export type ListAllConsultationsParams = {
status?: string;
};

export type ListCommunityPostsParams = {
limit?: number;
};