# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: M-ALDBANI platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [public]
      summary: List portfolio projects
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [public]
      summary: Get a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # PUBLIC - SERVICES
...
  /articles:
    get:
      operationId: listArticles
      tags: [public]
      summary: List articles
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [public]
      summary: Get an article by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

  # PUBLIC - CONTACT & BOOKING
  /contact:
    post:
      operationId: submitContact
      tags: [public]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /consultations/book:
    post:
      operationId: bookConsultation
      tags: [public]
      summary: Book a consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/slots:
    get:
      operationId: getAvailableSlots
      tags: [public]
      summary: Get available booking slots for a date
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # CLIENT PORTAL
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community posts
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}:
    get:
      operationId: getCommunityPost
      tags: [community]
      summary: Get a community post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/reactions:
    post:
      operationId: reactToPost
      tags: [community]
      summary: React to a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /community/posts/{id}/comments:
    post:
      operationId: commentOnPost
      tags: [community]
      summary: Comment on a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /community/posts/{id}/seen:
    post:
      operationId: markPostSeen
      tags: [community]
      summary: Mark post as seen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        image:
          type: string
        images:
          type: array
          items:
            type: string
        results:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
      required: [id, title, titleAr, description, descriptionAr, category, technologies, image, featured, order, createdAt]
...
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        titleAr:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        excerptAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        coverImage:
          type: string
        readTime:
          type: integer
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, titleAr, slug, excerpt, excerptAr, content, contentAr, category, coverImage, readTime, published, createdAt]
...
    Consultation:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: ["string", "null"]
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        type:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        meetingLink:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        paid:
          type: boolean
        createdAt:
          type: string
      required: [id, clientName, clientEmail, clientPhone, type, date, time, duration, status, paid, createdAt]
    BookingInput:
      type: object
      properties:
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        type:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        notes:
          type: string
      required: [clientName, clientEmail, clientPhone, type, date, time, duration]
...
    ContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        subject:
          type: string
      required: [name, email, message]
...
    CommunityPost:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
        image:
          type: ["string", "null"]
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        seenCount:
          type: integer
        publishedAt:
          type: string
      required: [id, title, content, type, reactions, comments, seenCount, publishedAt]
    CommunityPostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
        image:
          type: string
      required: [title, content, type]
    Reaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        userReacted:
          type: boolean
      required: [emoji, count, userReacted]
    ReactionInput:
      type: object
      properties:
        emoji:
          type: string
      required: [emoji]
    Comment:
      type: object
      properties:
        id:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, authorName, content, createdAt]
    CommentInput:
      type: object
      properties:
        content:
          type: string
      required: [content]